cnsn_dmcgb/requirements.txt
absl-py==0.15.0
cloudpickle==2.2.1
dm-env==1.5
dm-tree==0.1.7
einops==0.6.0
future==0.18.3
glfw==2.5.5
gym==0.25.2
gym-notices==0.0.8
imageio==2.9.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
kornia==0.5.11
lxml==4.9.2
numpy==1.19.5
opencv-python==4.6.0.66
packaging==23.0
#Pillow==6.2.0
PyOpenGL==3.1.6
scipy==1.7.3
six==1.16.0
termcolor==2.2.0
torchvision==0.8.2
tqdm==4.64.1
typing_extensions==4.4.0
xmltodict==0.13.0
zipp==3.12.0
